fix: 适配浮点数据

+++ filter_layer/utils.py
@@ -155,11 +155,11 @@ def get_user_seqs_and_max_item(data_file):
     item_set = set()
     for line in lines:
         items = line.strip().split()
-        items = [int(eval(item)) for item in items]
+        items = [float(eval(item)) for item in items]
         user_seq.append(items)
         item_set = item_set | set(items)
     max_item = max(item_set)
-    return user_seq, max_item
+    return user_seq, int(max_item)
 
 def get_user_seqs_and_sample(data_file, sample_file):
     lines = open(data_file).readlines()
@@ -167,7 +167,7 @@ def get_user_seqs_and_sample(data_file, sample_file):
     item_set = set()
     for line in lines:
         items = line.strip().split()
-        items = [int(eval(item)) for item in items]
+        items = [float(eval(item)) for item in items]
         user_seq.append(items)
         item_set = item_set | set(items)
     max_item = max(item_set)
@@ -176,12 +176,12 @@ def get_user_seqs_and_sample(data_file, sample_file):
     sample_seq = []
     for line in lines:
         items = line.strip().split()
-        items = [int(eval(item)) for item in items]
+        items = [float(eval(item)) for item in items]
         sample_seq.append(items)
 
     assert len(user_seq) == len(sample_seq)
 
-    return user_seq, max_item, num_users, sample_seq
+    return user_seq, int(max_item), num_users, sample_seq
 
 def get_user_seqs_and_sample4session_based(data_file, sample_file):
     lines = open(data_file).readlines()
@@ -189,7 +189,7 @@ def get_user_seqs_and_sample4session_based(data_file, sample_file):
     item_set = set()
     for line in lines:
         items = line.strip().split()
-        items = [int(eval(item)) for item in items]
+        items = [float(eval(item)) for item in items]
         user_seq.append(items)
         item_set = item_set | set(items)
     num_users = len(lines)
@@ -197,7 +197,7 @@ def get_user_seqs_and_sample4session_based(data_file, sample_file):
     sample_seq = []
     for line in lines:
         items = line.strip().split()
-        items = [int(eval(item)) for item in items]
+        items = [float(eval(item)) for item in items]
         sample_seq.append(items)
 
     return user_seq, num_users, sample_seq
